sport_backed/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Django==5.2.13
django-cors-headers==4.9.0
django-filter==25.2
django-stubs==6.0.2
django-stubs-ext==6.0.2
djangorestframework==3.17.1
djangorestframework-stubs==3.16.9
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
idna==3.11
inflection==0.5.1
packaging==26.0
pillow==12.2.0
pycparser==3.0
PyJWT==2.12.1
PyMySQL==1.1.2
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
sqlparse==0.5.5
types-PyYAML==6.0.12.20260408
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3